+++ src/pila.c
@@ -5,7 +5,7 @@
 
 struct _pila_t {
 	nodo_t *nodo_inicio;
-	int cant_nodos;
+	size_t cant_nodos;
 	nodo_t *nodo_fin;
 };
